refactor: 이벤트 도메인 패키지 이동
